requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Django==3.0.1
django-cors-headers==3.2.0
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
Pillow==7.0.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
